+++ packages/rocketchat-lib/lib/MessageTypes.js
@@ -1,24 +1,7 @@
 import { Meteor } from 'meteor/meteor';
+import { MessageTypes } from 'meteor/rocketchat:ui-utils';
 
-RocketChat.MessageTypes = new class {
-	constructor() {
-		this.types = {};
-	}
-
-	registerType(options) {
-		return this.types[options.id] = options;
-	}
-
-	getType(message) {
-		return this.types[message && message.t];
-	}
-
-	isSystemMessage(message) {
-		const type = this.types[message && message.t];
-		return type && type.system;
-	}
-
-};
+RocketChat.MessageTypes = MessageTypes;
 
 Meteor.startup(function() {
 	RocketChat.MessageTypes.registerType({
