+++ packages/rocketchat-ui-cached-collection/client/models/CachedCollection.js
@@ -340,19 +340,21 @@ export class CachedCollection {
 	}
 
 	async setupListener(eventType, eventName) {
-		const { Notifications } = await import('meteor/rocketchat:notifications');
-		const { RoomManager } = await import('meteor/rocketchat:ui');
-		Notifications[eventType || this.eventType](eventName || this.eventName, (t, record) => {
-			this.log('record received', t, record);
-			callbacks.run(`cachedCollection-received-${ this.name }`, record, t);
-			if (t === 'removed') {
-				this.collection.remove(record._id);
-				RoomManager.close(record.t + record.name);
-			} else {
-				this.collection.upsert({ _id: record._id }, _.omit(record, '_id'));
-			}
+		Meteor.startup(async() => {
+			const { Notifications } = await import('meteor/rocketchat:notifications');
+			const { RoomManager } = await import('meteor/rocketchat:ui');
+			Notifications[eventType || this.eventType](eventName || this.eventName, (t, record) => {
+				this.log('record received', t, record);
+				callbacks.run(`cachedCollection-received-${ this.name }`, record, t);
+				if (t === 'removed') {
+					this.collection.remove(record._id);
+					RoomManager.close(record.t + record.name);
+				} else {
+					this.collection.upsert({ _id: record._id }, _.omit(record, '_id'));
+				}
 
-			this.saveCache();
+				this.saveCache();
+			});
 		});
 	}
