+++ packages/rocketchat-ui-message/client/messageBox.js
@@ -499,7 +499,7 @@ Template.messageBox.events({
 		const timer = document.querySelector('.rc-message-box__timer');
 		const mic = document.querySelector('.rc-message-box__icon.mic');
 
-		chatMessages[RocketChat.openedRoom].recording = true;
+		chatMessages[RoomManager.openedRoom].recording = true;
 		AudioRecorder.start(function() {
 			const startTime = new Date;
 			timer.innerHTML = '00:00';
@@ -531,7 +531,7 @@ Template.messageBox.events({
 		}
 
 		AudioRecorder.stop();
-		chatMessages[RocketChat.openedRoom].recording = false;
+		chatMessages[RoomManager.openedRoom].recording = false;
 	},
 	'click .js-audio-message-check'(event) {
 		event.preventDefault();
@@ -547,12 +547,12 @@ Template.messageBox.events({
 			clearInterval(audioMessageIntervalId);
 		}
 
-		chatMessages[RocketChat.openedRoom].recording = false;
+		chatMessages[RoomManager.openedRoom].recording = false;
 		AudioRecorder.stop(function(blob) {
 
 			loader.classList.remove('active');
 			mic.classList.add('active');
-			const roomId = Session.get('openedRoom');
+			const roomId = Session.get('RoomManager.openedRoom');
 			const record = {
 				name: `${ TAPi18n.__('Audio record') }.mp3`,
 				size: blob.size,
@@ -658,7 +658,7 @@ Template.messageBox.events({
 			RocketChat.EmojiPicker.close();
 		} else {
 			RocketChat.EmojiPicker.open(event.currentTarget, (emoji) => {
-				const { input } = chatMessages[RocketChat.openedRoom];
+				const { input } = chatMessages[RoomManager.openedRoom];
 
 				const emojiValue = `:${ emoji }:`;
 
@@ -685,8 +685,8 @@ Template.messageBox.onRendered(function() {
 		const reply = $(input).data('reply');
 		self.dataReply.set(reply);
 	});
-	chatMessages[RocketChat.openedRoom] = chatMessages[RocketChat.openedRoom] || new ChatMessages;
-	chatMessages[RocketChat.openedRoom].input = this.$('.js-input-message').autogrow({
+	chatMessages[RoomManager.openedRoom] = chatMessages[RoomManager.openedRoom] || new ChatMessages;
+	chatMessages[RoomManager.openedRoom].input = this.$('.js-input-message').autogrow({
 		animate: true,
 		onInitialize: true,
 	}).on('autogrow', () => {
@@ -724,9 +724,9 @@ Meteor.startup(function() {
 	});
 	RocketChat.callbacks.add('enter-room', function() {
 		setTimeout(() => {
-			if (chatMessages[RocketChat.openedRoom].input) {
-				chatMessages[RocketChat.openedRoom].input.focus();
-				chatMessages[RocketChat.openedRoom].restoreText(RocketChat.openedRoom);
+			if (chatMessages[RoomManager.openedRoom].input) {
+				chatMessages[RoomManager.openedRoom].input.focus();
+				chatMessages[RoomManager.openedRoom].restoreText(RoomManager.openedRoom);
 			}
 		}, 200);
 	});
