+++ packages/rocketchat-ui-message/client/popup/messagePopupConfig.js
@@ -304,7 +304,7 @@ Template.messagePopupConfig.helpers({
 
 				// Get users from Server
 				if (items.length < 5 && filterText !== '') {
-					fetchUsersFromServerDelayed(filterText, items, cb, RocketChat.openedRoom);
+					fetchUsersFromServerDelayed(filterText, items, cb, RoomManager.openedRoom);
 				}
 
 				const all = {
@@ -365,7 +365,7 @@ Template.messagePopupConfig.helpers({
 				}).fetch();
 
 				if (records.length < 5 && filter && filter.trim() !== '') {
-					fetchRoomsFromServerDelayed(filter, records, cb, RocketChat.openedRoom);
+					fetchRoomsFromServerDelayed(filter, records, cb, RoomManager.openedRoom);
 				}
 				return records;
 			},
