+++ packages/rocketchat-ui-utils/lib/MessageProperties.js
@@ -1,0 +1,16 @@
+import { emoji } from 'meteor/rocketchat:emoji';
+import GraphemeSplitter from 'grapheme-splitter';
+
+const splitter = new GraphemeSplitter();
+
+export const messageProperties = {
+
+	length: ((message) => splitter.countGraphemes(message)),
+
+	messageWithoutEmojiShortnames: ((message) => message.replace(/:\w+:/gm, (match) => {
+		if (emoji.list[match] !== undefined) {
+			return ' ';
+		}
+		return match;
+	})),
+};
