+++ packages/rocketchat-ui-utils/tests/server.mocks.js
@@ -1,0 +1,8 @@
+import mock from 'mock-require';
+
+mock('meteor/rocketchat:emoji', {
+	emoji: {
+		list: {
+		},
+	},
+});
